+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/EmbeddedPaymentElementSnapshotTests.swift
@@ -50,9 +50,6 @@ class EmbeddedPaymentElementSnapshotTests: STPSnapshotTestCase, EmbeddedPaymentE
         XCTAssertTrue(delegateDidUpdateHeightCalled)
         XCTAssertFalse(delegateDidUpdatePaymentOptionCalled)
     }
-    
-    func _verifyView() {
-    }
 
     func testMarginsAreZero() async throws {
         let sut = try await EmbeddedPaymentElement.create(intentConfiguration: paymentIntentConfig, configuration: configuration)